ref(seer): Move organizationConfigIntegrationsQueryOptions to be co-located with its caller
